pyproject.toml
[project]
name = "tyr-agent"
version = "0.0.6"
description = "Agente LLM com execução de funções, histórico persistente, suporte a arquivos e orquestração de múltiplos agentes."
authors = [
  { name = "Witor Oliveira", email = "witoredson@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
  "google-generativeai",
  "openai",
  "python-dotenv"
]
keywords = ["llm", "agent", "gemini", "function-calling", "ai"]

[project.urls]
"Homepage" = "https://github.com/Drarlian/tyr-agent"
"Source" = "https://github.com/Drarlian/tyr-agent"
"Issues" = "https://github.com/Drarlian/tyr-agent/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"tyr_agent" = ["py.typed"]